% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_stats.R
\name{desc_stats}
\alias{desc_stats}
\title{Calculate Descriptive Statistics
Compute mean, variance, minimum, maximum values and range for a numeric vector}
\usage{
desc_stats(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
A descriptive statistics list containing mean, variance, min, max values and range
}
\description{
Calculate Descriptive Statistics
Compute mean, variance, minimum, maximum values and range for a numeric vector
}
\examples{
x<-c(1,2,3,4,5,6)
desc_stats(x)
}
